Improve Sign in with Google config
Remove unnecessary Sign in with Google iOS client ID from Dart code. The
ID is already in plist file.

Also add Dart file containing Sign in with Google backend client ID to
Git history. The ID is not a secret and it is already in a different
file which is Android project specific.