Fix missing WebSocket connection ping
Changing UI to conversation screen reset the timer which caused client
to not send ping to server and that caused server to timeout the
connection.

Now the 5 minute timer is always running and the 10 second timer is an
extra timer. It does not matter if the extra timer sends ping directly
after the first timer as this happens quite rarely.